Leader election options
