+++ app/app.go
@@ -23,8 +23,6 @@ import (
 	"k8s.io/client-go/kubernetes"
 	core_v1client "k8s.io/client-go/kubernetes/typed/core/v1"
 	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/leaderelection"
-	"k8s.io/client-go/tools/leaderelection/resourcelock"
 	"k8s.io/client-go/tools/record"
 	"k8s.io/client-go/util/flowcontrol"
 	"k8s.io/client-go/util/workqueue"
@@ -32,22 +30,8 @@ import (
 
 const (
 	defaultResyncPeriod  = 20 * time.Minute
-	defaultLeaseDuration = 15 * time.Second
-	defaultRenewDeadline = 10 * time.Second
-	defaultRetryPeriod   = 2 * time.Second
 	defaultAuxServerAddr = ":9090"
 )
-
-// See kubernetes/kubernetes/pkg/apis/componentconfig/types.go LeaderElectionConfiguration
-// for leader election configuration description.
-type LeaderElectionConfig struct {
-	LeaderElect        bool
-	LeaseDuration      time.Duration
-	RenewDeadline      time.Duration
-	RetryPeriod        time.Duration
-	ConfigMapNamespace string
-	ConfigMapName      string
-}
 
 type PrometheusRegistry interface {
 	prometheus.Registerer
@@ -61,15 +45,15 @@ type App struct {
 
 	// Name is the name of the application. It must only contain alphanumeric
 	// characters.
-	Name                 string
-	RestConfig           *rest.Config
-	ResyncPeriod         time.Duration
-	Namespace            string
-	Controllers          []ctrl.Constructor
-	Workers              int
-	LeaderElectionConfig LeaderElectionConfig
-	AuxListenOn          string
-	Debug                bool
+	Name                  string
+	RestConfig            *rest.Config
+	ResyncPeriod          time.Duration
+	Namespace             string
+	Controllers           []ctrl.Constructor
+	Workers               int
+	LeaderElectionOptions LeaderElectionOptions
+	AuxListenOn           string
+	Debug                 bool
 }
 
 func (a *App) Run(ctx context.Context) (retErr error) {
@@ -142,66 +126,14 @@ func (a *App) Run(ctx context.Context) (retErr error) {
 	})
 
 	// Leader election
-	if a.LeaderElectionConfig.LeaderElect {
-		a.Logger.Info("Starting leader election", logz.NamespaceName(a.LeaderElectionConfig.ConfigMapNamespace))
-		ctx, err = DoLeaderElection(ctx, a.Logger, a.Name, a.LeaderElectionConfig, a.MainClient.CoreV1(), recorder)
+	if a.LeaderElectionOptions.LeaderElect {
+		a.Logger.Info("Starting leader election", logz.NamespaceName(a.LeaderElectionOptions.ConfigMapNamespace))
+		ctx, err = DoLeaderElection(ctx, a.Logger, a.Name, a.LeaderElectionOptions, a.MainClient.CoreV1(), recorder)
 		if err != nil {
 			return err
 		}
 	}
 	return generic.Run(ctx)
-}
-
-// DoLeaderElection starts leader election and blocks until it acquires the lease.
-// Returned context is cancelled once the lease is lost or ctx signals done.
-func DoLeaderElection(ctx context.Context, logger *zap.Logger, component string, config LeaderElectionConfig, configMapsGetter core_v1client.ConfigMapsGetter, recorder record.EventRecorder) (context.Context, error) {
-	id, err := os.Hostname()
-	if err != nil {
-		return nil, err
-	}
-	ctxRet, cancel := context.WithCancel(ctx)
-	startedLeading := make(chan struct{})
-	le, err := leaderelection.NewLeaderElector(leaderelection.LeaderElectionConfig{
-		Lock: &resourcelock.ConfigMapLock{
-			ConfigMapMeta: meta_v1.ObjectMeta{
-				Namespace: config.ConfigMapNamespace,
-				Name:      config.ConfigMapName,
-			},
-			Client: configMapsGetter,
-			LockConfig: resourcelock.ResourceLockConfig{
-				Identity:      id + "-" + component,
-				EventRecorder: recorder,
-			},
-		},
-		LeaseDuration: config.LeaseDuration,
-		RenewDeadline: config.RenewDeadline,
-		RetryPeriod:   config.RetryPeriod,
-		Callbacks: leaderelection.LeaderCallbacks{
-			OnStartedLeading: func(ctx context.Context) {
-				logger.Info("Started leading")
-				close(startedLeading)
-			},
-			OnStoppedLeading: func() {
-				logger.Info("Leader status lost")
-				cancel()
-			},
-		},
-	})
-	if err != nil {
-		cancel()
-		return nil, err
-	}
-	go func() {
-		// note: because le.Run() also adds a logging panic handler panics with be logged 3 times
-		defer logz.LogStructuredPanic()
-		le.Run(ctx)
-	}()
-	select {
-	case <-ctx.Done():
-		return nil, ctx.Err()
-	case <-startedLeading:
-		return ctxRet, nil
-	}
 }
 
 // CancelOnInterrupt calls f when os.Interrupt or SIGTERM is received.
@@ -234,28 +166,7 @@ func NewFromFlags(name string, controllers []ctrl.Constructor, flagset *flag.Fla
 	flagset.StringVar(&a.AuxListenOn, "aux-listen-on", defaultAuxServerAddr, "Auxiliary address to listen on. Used for Prometheus metrics server and pprof endpoint. Empty to disable")
 	qps := flagset.Float64("api-qps", 5, "Maximum queries per second when talking to Kubernetes API")
 
-	// This flag is off by default only because leader election package says it is ALPHA API.
-	flagset.BoolVar(&a.LeaderElectionConfig.LeaderElect, "leader-elect", false, ""+
-		"Start a leader election client and gain leadership before "+
-		"executing the main loop. Enable this when running replicated "+
-		"components for high availability")
-	flagset.DurationVar(&a.LeaderElectionConfig.LeaseDuration, "leader-elect-lease-duration", defaultLeaseDuration, ""+
-		"The duration that non-leader candidates will wait after observing a leadership "+
-		"renewal until attempting to acquire leadership of a led but unrenewed leader "+
-		"slot. This is effectively the maximum duration that a leader can be stopped "+
-		"before it is replaced by another candidate. This is only applicable if leader "+
-		"election is enabled")
-	flagset.DurationVar(&a.LeaderElectionConfig.RenewDeadline, "leader-elect-renew-deadline", defaultRenewDeadline, ""+
-		"The interval between attempts by the acting master to renew a leadership slot "+
-		"before it stops leading. This must be less than or equal to the lease duration. "+
-		"This is only applicable if leader election is enabled")
-	flagset.DurationVar(&a.LeaderElectionConfig.RetryPeriod, "leader-elect-retry-period", defaultRetryPeriod, ""+
-		"The duration the clients should wait between attempting acquisition and renewal "+
-		"of a leadership. This is only applicable if leader election is enabled")
-	flagset.StringVar(&a.LeaderElectionConfig.ConfigMapNamespace, "leader-elect-configmap-namespace", meta_v1.NamespaceDefault,
-		"Namespace to use for leader election ConfigMap. This is only applicable if leader election is enabled")
-	flagset.StringVar(&a.LeaderElectionConfig.ConfigMapName, "leader-elect-configmap-name", name+"-leader-elect",
-		"ConfigMap name to use for leader election. This is only applicable if leader election is enabled")
+	BindLeaderElectionFlags(name, &a.LeaderElectionOptions, flagset)
 	configFileFrom := flagset.String("client-config-from", "in-cluster",
 		"Source of REST client configuration. 'in-cluster' (default), 'environment' and 'file' are valid options.")
 	configFileName := flagset.String("client-config-file-name", "",
